src/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
ipykernel==6.29.4
pydantic-settings==2.2.1
pillow==10.3.0
dill==0.3.8
gdown==5.2.0
google-generativeai==0.5.4

